Remove use of oteltest in otelbeego
