+++ .github/dependabot.yml
@@ -137,6 +137,16 @@ updates:
     schedule:
       interval: "weekly"
       day: "sunday"
+  -
+    package-ecosystem: "gomod"
+    directory: "/instrumentation/github.com/astaxie/beego/otelbeego/test"
+    labels:
+      - dependencies
+      - go
+      - "Skip Changelog"
+    schedule:
+      interval: "weekly"
+      day: "sunday"
   - package-ecosystem: "gomod"
     directory: "/instrumentation/github.com/aws/aws-sdk-go-v2/otelaws"
     labels:
